+++ netpyne/cell.py
@@ -7,6 +7,7 @@ Contributors: salvadordura@gmail.com
 """
 
 from numbers import Number
+from copy import deepcopy
 from neuron import h # Import NEURON
 from specs import Dict
 import sim
@@ -103,6 +104,13 @@ class Cell (object):
                         sec['ions'][ionName] = Dict()  
                     for ionParamName,ionParamValue in ionParams.iteritems():  # add params of the ion
                         sec['ions'][ionName][ionParamName] = ionParamValue
+
+
+            # add synMechs
+            if 'synMechs' in sectParams:
+                for synMech in sectParams['synMechs']:
+                    if 'label' in synMech and 'loc' in synMech:
+                        self.addSynMech(synLabel=synMech['label'], secLabel=sectName, loc=synMech['loc'])
 
             # add point processes
             if 'pointps' in sectParams:
@@ -339,6 +347,8 @@ class Cell (object):
                 synMech = next((synMech for synMech in sec['synMechs'] if synMech['label']==synLabel and synMech['loc']==loc), None)
                 if not synMech:  # if synMech not in section, then create
                     synMech = Dict({'label': synLabel, 'loc': loc})
+                    for paramName, paramValue in synMechParams.iteritems():
+                        synMech[paramName] = paramValue
                     sec['synMechs'].append(synMech)
 
             if sim.cfg.createNEURONObj:
@@ -357,13 +367,110 @@ class Cell (object):
                         if synParamName not in ['label', 'mod', 'selfNetCon', 'loc']:
                             setattr(synMech['hSyn'], synParamName, synParamValue)
                         elif synParamName == 'selfNetCon':  # create self netcon required for some synapses (eg. homeostatic)
-                            synMech['hNetCon'] = h.NetCon(sec['hSec'](loc)._ref_v, synMech['hSyn'], sec=sec['hSec'])
+                            secLabelNetCon = synParamValue.get('sec', 'soma')
+                            locNetCon = synParamValue.get('loc', 0.5)
+                            secNetCon = self.secs.get(secLabelNetCon, None)
+                            synMech['hNetCon'] = h.NetCon(secNetCon['hSec'](locNetCon)._ref_v, synMech['hSyn'], sec=secNetCon['hSec'])
                             for paramName,paramValue in synParamValue.iteritems():
                                 if paramName == 'weight':
                                     synMech['hNetCon'].weight[0] = paramValue
-                                else:
+                                elif paramName not in ['sec', 'loc']:
                                     setattr(synMech['hNetCon'], paramName, paramValue)
             return synMech
+
+
+    def modifySynMechs (self, params):
+        conditionsMet = 1
+        if 'cellConds' in params:
+            if conditionsMet:
+                for (condKey,condVal) in params['cellConds'].iteritems():  # check if all conditions are met
+                    # check if conditions met
+                    if isinstance(condVal, list):
+                        if self.tags.get(condKey) < condVal[0] or self.tags.get(condKey) > condVal[1]:
+                            conditionsMet = 0
+                            break
+                    elif self.tags.get(condKey) != condVal: 
+                        conditionsMet = 0
+                        break
+
+        if conditionsMet:
+            for secLabel,sec in self.secs.iteritems():
+                for synMech in sec['synMechs']:
+                    conditionsMet = 1
+                    if 'conds' in params:
+                        for (condKey,condVal) in params['conds'].iteritems():  # check if all conditions are met
+                            # check if conditions met
+                            if condKey == 'sec':
+                                if condVal != secLabel:
+                                    conditionsMet = 0
+                                    break
+                            elif isinstance(condVal, list) and isinstance(condVal[0], Number):
+                                if synMech.get(condKey) < condVal[0] or synMech.get(condKey) > condVal[1]:
+                                    conditionsMet = 0
+                                    break
+                            elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                                if synMech.get(condKey) not in condVal:
+                                    conditionsMet = 0
+                                    break 
+                            elif synMech.get(condKey) != condVal: 
+                                conditionsMet = 0
+                                break
+
+                    if conditionsMet:  # if all conditions are met, set values for this cell
+                        exclude = ['conds', 'cellConds', 'label', 'mod', 'selfNetCon', 'loc']
+                        for synParamName,synParamValue in {k: v for k,v in params.iteritems() if k not in exclude}.iteritems():
+                            if sim.cfg.createPyStruct: 
+                                synMech[synParamName] = synParamValue
+                            if sim.cfg.createNEURONObj:
+                                try: 
+                                    setattr(synMech['hSyn'], synParamName, synParamValue)
+                                except:
+                                    print 'Error setting %s=%s on synMech' % (synParamName, str(synParamValue))
+
+
+    
+    
+    # Custom code for time-dependently shaping the weight of a NetCon corresponding to a NetStim.
+    def _shapeStim(self, isi=1, variation=0, width=0.05, weight=10, start=0, finish=1, stimshape='gaussian'):
+        from pylab import r_, convolve, shape, exp, zeros, hstack, array, rand
+        
+        # Create event times
+        timeres = 0.001 # Time resolution = 1 ms = 500 Hz (DJK to CK: 500...?)
+        pulselength = 10 # Length of pulse in units of width
+        currenttime = 0
+        timewindow = finish-start
+        allpts = int(timewindow/timeres)
+        output = []
+        while currenttime<timewindow:
+            # Note: The timeres/2 subtraction acts as an eps to avoid later int rounding errors.
+            if currenttime>=0 and currenttime<timewindow-timeres/2: output.append(currenttime)
+            currenttime = currenttime+isi+variation*(rand()-0.5)
+        
+        # Create single pulse
+        npts = pulselength*width/timeres
+        x = (r_[0:npts]-npts/2+1)*timeres
+        if stimshape=='gaussian': 
+            pulse = exp(-2*(2*x/width-1)**2) # Offset by 2 standard deviations from start
+            pulse = pulse/max(pulse)
+        elif stimshape=='square': 
+            pulse = zeros(shape(x))
+            pulse[int(npts/2):int(npts/2)+int(width/timeres)] = 1 # Start exactly on time
+        else:
+            raise Exception('Stimulus shape "%s" not recognized' % stimshape)
+        
+        # Create full stimulus
+        events = zeros((allpts))
+        events[array(array(output)/timeres,dtype=int)] = 1
+        fulloutput = convolve(events,pulse,mode='full')*weight # Calculate the convolved input signal, scaled by rate
+        fulloutput = fulloutput[npts/2-1:-npts/2]   # Slices out where the convolved pulse train extends before and after sequence of allpts.
+        fulltime = (r_[0:allpts]*timeres+start)*1e3 # Create time vector and convert to ms
+        
+        fulltime = hstack((0,fulltime,fulltime[-1]+timeres*1e3)) # Create "bookends" so always starts and finishes at zero
+        fulloutput = hstack((0,fulloutput,0)) # Set weight to zero at either end of the stimulus period
+        events = hstack((0,events,0)) # Ditto
+        stimvecs = deepcopy([fulltime, fulloutput, events]) # Combine vectors into a matrix                   
+        
+        return stimvecs  
 
 
     def addConn (self, params, netStimParams = None):
@@ -441,6 +548,40 @@ class Cell (object):
                 netcon.threshold = params['threshold']  # set Netcon threshold
                 self.conns[-1]['hNetcon'] = netcon  # add netcon object to dict in conns list
             
+
+                # Add time-dependent weight shaping
+                if 'shape' in params and params['shape']:
+                    
+                    temptimevecs = []
+                    tempweightvecs = []
+                    
+                    # Default shape
+                    pulsetype = params['shape']['pulseType'] if 'pulseType' in params['shape'] else 'square'
+                    pulsewidth = params['shape']['pulseWidth'] if 'pulseWidth' in params['shape'] else 100.0
+                    pulseperiod = params['shape']['pulsePeriod'] if 'pulsePeriod' in params['shape'] else 100.0
+                    
+                    # Determine on-off switching time pairs for stimulus, where default is always on
+                    if 'switchOnOff' not in params['shape']:
+                        switchtimes = [0, sim.cfg.duration]
+                    else:
+                        if not params['shape']['switchOnOff'] == sorted(params['shape']['switchOnOff']):
+                            raise Exception('On-off switching times for a particular stimulus are not monotonic')   
+                        switchtimes = deepcopy(params['shape']['switchOnOff'])
+                        switchtimes.append(sim.cfg.duration)
+                    
+                    switchiter = iter(switchtimes)
+                    switchpairs = zip(switchiter,switchiter)
+                    for pair in switchpairs:
+                        # Note: Cliff's makestim code is in seconds, so conversions from ms to s occurs in the args.
+                        stimvecs = self._shapeStim(width=float(pulsewidth)/1000.0, isi=float(pulseperiod)/1000.0, weight=params['weight'], start=float(pair[0])/1000.0, finish=float(pair[1])/1000.0, stimshape=pulsetype)
+                        temptimevecs.extend(stimvecs[0])
+                        tempweightvecs.extend(stimvecs[1])
+                    
+                    self.conns[-1]['shapeTimeVec'] = h.Vector().from_python(temptimevecs)
+                    self.conns[-1]['shapeWeightVec'] = h.Vector().from_python(tempweightvecs)
+                    self.conns[-1]['shapeWeightVec'].play(netcon._ref_weight[weightIndex], self.conns[-1]['shapeTimeVec'])
+
+
                 # Add plasticity
                 self._addConnPlasticity(params, sec, netcon, weightIndex)
 
@@ -470,7 +611,7 @@ class Cell (object):
                             conditionsMet = 0
                             break
                     elif isinstance(condVal, list) and isinstance(condVal[0], str):
-                        if self.tags[condKey] not in condVal:
+                        if compareTo not in condVal:
                             conditionsMet = 0
                             break 
                     elif compareTo != condVal: 
@@ -511,57 +652,59 @@ class Cell (object):
 
 
     def modifyStims (self, params):
-        for stim in self.stims:
-            conditionsMet = 1
-
-            if 'conds' in params:
-                for (condKey,condVal) in params['conds'].iteritems():  # check if all conditions are met
+        conditionsMet = 1
+        if 'cellConds' in params:
+            if conditionsMet:
+                for (condKey,condVal) in params['cellConds'].iteritems():  # check if all conditions are met
                     # check if conditions met
-                    if isinstance(condVal, list) and isinstance(condVal[0], Number):
-                        if stim.get(condKey) < condVal[0] or stim.get(condKey) > condVal[1]:
+                    if isinstance(condVal, list):
+                        if self.tags.get(condKey) < condVal[0] or self.tags.get(condKey) > condVal[1]:
                             conditionsMet = 0
                             break
-                    elif isinstance(condVal, list) and isinstance(condVal[0], str):
-                        if self.tags[condKey] not in condVal:
-                            conditionsMet = 0
-                            break 
-                    elif stim[condKey] != condVal: 
+                    elif self.tags.get(condKey) != condVal: 
                         conditionsMet = 0
                         break
 
-            if conditionsMet and 'cellConds' in params:
-                if conditionsMet:
-                    for (condKey,condVal) in params['cellConds'].iteritems():  # check if all conditions are met
+        if conditionsMet == 1:
+            for stim in self.stims:
+                conditionsMet = 1
+
+                if 'conds' in params:
+                    for (condKey,condVal) in params['conds'].iteritems():  # check if all conditions are met
                         # check if conditions met
-                        if isinstance(condVal, list):
-                            if self.tags.get(condKey) < condVal[0] or self.tags.get(condKey) > condVal[1]:
+                        if isinstance(condVal, list) and isinstance(condVal[0], Number):
+                            if stim.get(condKey) < condVal[0] or stim.get(condKey) > condVal[1]:
                                 conditionsMet = 0
                                 break
-                        elif self.tags.get(condKey) != condVal: 
+                        elif isinstance(condVal, list) and isinstance(condVal[0], str):
+                            if stim.get(condKey) not in condVal:
+                                conditionsMet = 0
+                                break 
+                        elif stim.get(condKey) != condVal: 
                             conditionsMet = 0
                             break
 
-            if conditionsMet:  # if all conditions are met, set values for this cell
-                if stim['type'] == 'NetStim':  # for netstims, find associated netcon
-                    conn = next((conn for conn in self.conns if conn['source'] == stim['source']), None)
-                if sim.cfg.createPyStruct:
-                    for paramName, paramValue in {k: v for k,v in params.iteritems() if k not in ['conds','cellConds']}.iteritems():
-                        if stim['type'] == 'NetStim' and paramName in ['weight', 'delay', 'threshold']:
-                            conn[paramName] = paramValue
-                        else:
-                            stim[paramName] = paramValue
-                if sim.cfg.createNEURONObj:
-                    for paramName, paramValue in {k: v for k,v in params.iteritems() if k not in ['conds','cellConds']}.iteritems():
-                        try:
-                            if stim['type'] == 'NetStim':
-                                if paramName == 'weight':
-                                    conn['hNetcon'].weight[0] = paramValue
-                                elif paramName in ['delay', 'threshold']:
-                                    setattr(conn['hNetcon'], paramName, paramValue)
+                if conditionsMet:  # if all conditions are met, set values for this cell
+                    if stim['type'] == 'NetStim':  # for netstims, find associated netcon
+                        conn = next((conn for conn in self.conns if conn['source'] == stim['source']), None)
+                    if sim.cfg.createPyStruct:
+                        for paramName, paramValue in {k: v for k,v in params.iteritems() if k not in ['conds','cellConds']}.iteritems():
+                            if stim['type'] == 'NetStim' and paramName in ['weight', 'delay', 'threshold']:
+                                conn[paramName] = paramValue
                             else:
-                                setattr(stim['h'+stim['type']], paramName, paramValue)
-                        except:
-                            print 'Error setting %s=%s on stim' % (paramName, str(paramValue))
+                                stim[paramName] = paramValue
+                    if sim.cfg.createNEURONObj:
+                        for paramName, paramValue in {k: v for k,v in params.iteritems() if k not in ['conds','cellConds']}.iteritems():
+                            try:
+                                if stim['type'] == 'NetStim':
+                                    if paramName == 'weight':
+                                        conn['hNetcon'].weight[0] = paramValue
+                                    elif paramName in ['delay', 'threshold']:
+                                        setattr(conn['hNetcon'], paramName, paramValue)
+                                else:
+                                    setattr(stim['h'+stim['type']], paramName, paramValue)
+                            except:
+                                print 'Error setting %s=%s on stim' % (paramName, str(paramValue))
 
 
     def addNetStim (self, params, stimContainer=None):
@@ -627,6 +770,7 @@ class Cell (object):
                 'delay': params.get('delay'),
                 'threshold': params.get('threshold'),
                 'synsPerConn': params.get('synsPerConn'),
+                'shape': params.get('shape'),
                 'plast': params.get('plast')}
 
             netStimParams = {'source': params['source'],
@@ -851,14 +995,34 @@ class Cell (object):
                             ptr = synMech['hSyn'].__getattribute__('_ref_'+params['var'])
                         else:  # eg. soma(0.5)._ref_v
                             ptr = self.secs[params['sec']]['hSec'](params['loc']).__getattribute__('_ref_'+params['var'])
+                    elif 'synMech' in params:  # special case where want to record from multiple synMechs
+                        if 'sec' in params:
+                            sec = self.secs[params['sec']]
+                            synMechs = [synMech for synMech in sec['synMechs'] if synMech['label']==params['synMech']]
+                            ptr = [synMech['hSyn'].__getattribute__('_ref_'+params['var']) for synMech in synMechs]
+                            secLocs = [params.sec+str(synMech['loc']) for synMech in synMechs]
+                        else: 
+                            ptr = []
+                            secLocs = []
+                            for secName,sec in self.secs.iteritems():
+                                synMechs = [synMech for synMech in sec['synMechs'] if synMech['label']==params['synMech']]
+                                ptr.extend([synMech['hSyn'].__getattribute__('_ref_'+params['var']) for synMech in synMechs])
+                                secLocs.extend([secName+'_'+str(synMech['loc']) for synMech in synMechs])
+
                     else:
                         if 'pointp' in params: # eg. soma.izh._ref_u
                             if params['pointp'] in self.secs[params['sec']]['pointps']:
                                 ptr = self.secs[params['sec']]['pointps'][params['pointp']]['hPointp'].__getattribute__('_ref_'+params['var'])
 
                     if ptr:  # if pointer has been created, then setup recording
-                        sim.simData[key]['cell_'+str(self.gid)] = h.Vector(sim.cfg.duration/sim.cfg.recordStep+1).resize(0)
-                        sim.simData[key]['cell_'+str(self.gid)].record(ptr, sim.cfg.recordStep)
+                        if isinstance(ptr, list):
+                            sim.simData[key]['cell_'+str(self.gid)] = {}
+                            for ptrItem,secLoc in zip(ptr, secLocs):
+                                sim.simData[key]['cell_'+str(self.gid)][secLoc] = h.Vector(sim.cfg.duration/sim.cfg.recordStep+1).resize(0)
+                                sim.simData[key]['cell_'+str(self.gid)][secLoc].record(ptrItem, sim.cfg.recordStep)
+                        else:
+                            sim.simData[key]['cell_'+str(self.gid)] = h.Vector(sim.cfg.duration/sim.cfg.recordStep+1).resize(0)
+                            sim.simData[key]['cell_'+str(self.gid)].record(ptr, sim.cfg.recordStep)
                         if sim.cfg.verbose: print '  Recording ', key, 'from cell ', self.gid, ' with parameters: ',str(params)
                 except:
                     if sim.cfg.verbose: print '  Cannot record ', key, 'from cell ', self.gid